# src/domain/post/post.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Patch,
  Param,
  Delete,
  UseGuards,
} from "@nestjs/common";
import { PostService } from "./post.service";
import { CreatePostDto } from "./dto/create-post.dto";
import { UpdatePostDto } from "./dto/update-post.dto";
import { FirebaseAuthGuard } from "src/services/auth/firebase/firebase-auth.guard";
import {ApiTags} from "@nestjs/swagger";

@ApiTags('post')
@Controller("post")
// @UseGuards(FirebaseAuthGuard)
export class PostController {
  constructor(private readonly postService: PostService) {}

  @Post("add")
  create(@Body() createPostDto: CreatePostDto) {
    return this.postService.create(createPostDto);
  }

  @Get("all/:id")
  findAll(@Param("id") id: string) {
    return this.postService.findAll(id);
  }

  @Get(":id")
  findOne(@Param("id") id: string) {
    return this.postService.findOne(+id);
  }

  @Get("user/:id")
  findAllByUserId(@Param("id") id: string) {
    return this.postService.findAllByUserId(id);
  }

  @Patch(":id")
  update(@Param("id") id: string, @Body() updatePostDto: UpdatePostDto) {
    return this.postService.update(+id, updatePostDto);
  }

  @Delete(":id")
  remove(@Param("id") id: string) {
    return this.postService.remove(+id);
  }
}